--- a/test/performance/vectorize.cpp
+++ b/test/performance/vectorize.cpp
@@ -28,7 +28,7 @@ bool test() {
     const Target target = get_jit_target_from_environment();
     const int vec_width = target.natural_vector_size<A>();
 
-    int W = vec_width * 1;
+    int W = vec_width * 2 - 1;
     int H = 10000;
 
     Buffer<A> input(W, H + 20);
@@ -38,55 +38,57 @@ bool test() {
         }
     }
 
-    Var x, y;
-    Func f, g;
+    for (TailStrategy tail_strategy : {TailStrategy::ShiftInwards, TailStrategy::GuardWithIf}) {
+        Var x, y;
+        Func f, g;
 
-    Expr e = input(x, y);
-    for (int i = 1; i < 5; i++) {
-        e = e + input(x, y + i);
-    }
+        Expr e = input(x, y);
+        for (int i = 1; i < 5; i++) {
+            e = e + input(x, y + i);
+        }
 
-    for (int i = 5; i >= 0; i--) {
-        e = e + input(x, y + i);
-    }
+        for (int i = 5; i >= 0; i--) {
+            e = e + input(x, y + i);
+        }
 
-    f(x, y) = e;
-    g(x, y) = e;
-    f.bound(x, 0, vec_width).vectorize(x);
+        f(x, y) = e;
+        g(x, y) = e;
+        f.vectorize(x, vec_width, tail_strategy);
 
-    // Stop llvm from auto-vectorizing the scalar case and messing up
-    // the comparison. Also causes cache effects, but the entire input
-    // is small enough to fit in cache.
-    g.reorder(y, x);
+        // Stop llvm from auto-vectorizing the scalar case and messing up
+        // the comparison. Also causes cache effects, but the entire input
+        // is small enough to fit in cache.
+        g.reorder(y, x);
 
-    Buffer<A> outputg = g.realize({W, H});
-    Buffer<A> outputf = f.realize({W, H});
+        Buffer<A> outputg = g.realize({W, H});
+        Buffer<A> outputf = f.realize({W, H});
 
-    double t_g = benchmark([&]() {
-        g.realize(outputg);
-    });
-    double t_f = benchmark([&]() {
-        f.realize(outputf);
-    });
+        double t_g = benchmark([&]() {
+            g.realize(outputg);
+        });
+        double t_f = benchmark([&]() {
+            f.realize(outputf);
+        });
 
-    for (int y = 0; y < H; y++) {
-        for (int x = 0; x < W; x++) {
-            if (outputf(x, y) != outputg(x, y)) {
-                printf("%s x %d failed at %d %d: %d vs %d\n",
-                       string_of_type<A>(), vec_width,
-                       x, y,
-                       (int)outputf(x, y),
-                       (int)outputg(x, y));
-                return false;
+        for (int y = 0; y < H; y++) {
+            for (int x = 0; x < W; x++) {
+                if (outputf(x, y) != outputg(x, y)) {
+                    printf("%s x %d failed at %d %d: %d vs %d\n",
+                           string_of_type<A>(), vec_width,
+                           x, y,
+                           (int)outputf(x, y),
+                           (int)outputg(x, y));
+                    return false;
+                }
             }
         }
-    }
 
-    printf("Vectorized vs scalar (%s x %d): %1.3gms %1.3gms. Speedup = %1.3f\n",
-           string_of_type<A>(), vec_width, t_f * 1e3, t_g * 1e3, t_g / t_f);
+        printf("Vectorized vs scalar (%s x %d): %1.3gms %1.3gms. Speedup = %1.3f\n",
+               string_of_type<A>(), vec_width, t_f * 1e3, t_g * 1e3, t_g / t_f);
 
-    if (t_f > t_g) {
-        return false;
+        if (t_f > t_g) {
+            return false;
+        }
     }
 
     return true;
@@ -102,14 +104,14 @@ int main(int argc, char **argv) {
     bool ok = true;
 
     // Only native vector widths for now
-    ok = ok && test<float>();
-    ok = ok && test<double>();
     ok = ok && test<uint8_t>();
     ok = ok && test<int8_t>();
     ok = ok && test<uint16_t>();
     ok = ok && test<int16_t>();
     ok = ok && test<uint32_t>();
     ok = ok && test<int32_t>();
+    ok = ok && test<float>();
+    ok = ok && test<double>();
 
     if (!ok) {
         return -1;