@dsharletg 2021-03-26
Closing in favor of #5856. We can't assume predicated loads/stores are always faster.